Update num-rational and num-bigint to 0.4.0.
Use i64-based rational numbers rather than isize-based.

+++ src/css/value.rs
@@ -127,7 +127,7 @@ impl Value {
 
     /// Check that this value is an integer.
     #[deprecated]
-    pub fn integer_value(&self) -> Result<isize, Error> {
+    pub fn integer_value(&self) -> Result<i64, Error> {
         match self {
             &Value::Numeric(ref num, ..) => num
                 .value
